+++ cypress/locators/button/index.js
@@ -3,10 +3,6 @@ import {
   BUTTON_DATA_COMPONENT_PREVIEW,
 } from "./locators";
 
-// component preview locators
-export const buttonDataComponentIFrame = () =>
-  cy.iFrame(BUTTON_DATA_COMPONENT_PREVIEW);
-
 // component preview locators in no IFrame
 export const buttonDataComponent = () => cy.get(BUTTON_DATA_COMPONENT_PREVIEW);
 export const buttonSubtextPreview = () => cy.get(BUTTON_SUBTEXT_PREVIEW);