+++ cypress/support/step-definitions/grouped-character-steps.js
@@ -1,16 +1,16 @@
 import {
   getKnobsInput,
-  commonDataElementInputPreviewNoIframe,
+  commonDataElementInputPreview,
 } from "../../locators";
 
 When("I put {string} example grouped character in no Iframe", (text) => {
-  commonDataElementInputPreviewNoIframe()
+  commonDataElementInputPreview()
     .clear()
     .type(text, { delay: 1000, force: true });
 });
 
 Then("example grouped character is {string}", (text) => {
-  commonDataElementInputPreviewNoIframe().should("have.value", text);
+  commonDataElementInputPreview().should("have.value", text);
 });
 
 When(