+++ src/components/form/form.spec.js
@@ -26,6 +26,8 @@ import Button from "../button";
 import { FieldsetStyle } from "../fieldset/fieldset.style";
 import StyledSearch from "../search/search.style";
 
+jest.mock("lodash/debounce", () => jest.fn((fn) => fn));
+
 describe("Form", () => {
   let wrapper;
 
@@ -492,6 +494,12 @@ describe("Form", () => {
           }
         }
       );
+
+      it("when there are no errors and warnings", () => {
+        wrapper.setProps({ errorCount: 0, warningCount: 0 });
+
+        expect(wrapper.find(StyledInternalSummary).exists()).toBe(false);
+      });
     });
   });