+++ src/components/pod/pod.stories.js
@@ -27,9 +27,11 @@ export const Default = () => {
   const alignTitle = title
     ? select("alignTitle", POD_ALIGNMENTS, Pod.defaultProps.alignTitle)
     : undefined;
-  const description = text("description", "");
   const footer = text("footer", "");
   const onEdit = boolean("onEdit", false);
+  const onDelete = boolean("onDelete", false);
+  const onUndo = boolean("onUndo", false);
+  const softDelete = boolean("softDelete", false);
   const editContentFullWidth = onEdit
     ? boolean("editContentFullWidth", false)
     : undefined;
@@ -51,9 +53,11 @@ export const Default = () => {
       title={title}
       subtitle={subtitle}
       alignTitle={alignTitle}
-      description={description}
       footer={footer}
       onEdit={onEdit ? action("edit") : undefined}
+      onDelete={onDelete ? action("delete") : undefined}
+      onUndo={onUndo ? action("undo") : undefined}
+      softDelete={softDelete}
       editContentFullWidth={editContentFullWidth}
       displayEditButtonOnHover={displayEditButtonOnHover}
       triggerEditOnContent={triggerEditOnContent}