+++ src/components/select/select-textbox/select-textbox.component.js
@@ -1,7 +1,7 @@
 import React from "react";
 import PropTypes from "prop-types";
-import I18n from "i18n-js";
 import Textbox from "../../textbox";
+import useLocale from "../../../hooks/__internal__/useLocale";
 
 const SelectTextbox = ({
   value,
@@ -16,9 +16,7 @@ const SelectTextbox = ({
   required,
   ...restProps
 }) => {
-  const defaultPlaceholder = I18n.t("select.placeholder", {
-    defaultValue: "Please Select...",
-  });
+  const l = useLocale();
 
   function handleTextboxClick(event) {
     if (disabled || readOnly) {
@@ -46,7 +44,7 @@ const SelectTextbox = ({
 
   function getTextboxProps() {
     return {
-      placeholder: placeholder || defaultPlaceholder,
+      placeholder: placeholder || l.select.placeholder(),
       disabled,
       readOnly,
       required,