+++ src/components/textarea/textarea.component.js
@@ -1,6 +1,5 @@
 import React from "react";
 import PropTypes from "prop-types";
-import I18n from "i18n-js";
 import styledSystemPropTypes from "@styled-system/prop-types";
 
 import { InputPresentation } from "../../__internal__/input";
@@ -9,13 +8,13 @@ import CharacterCount from "./character-count";
 import Input from "../../__internal__/input/input.component";
 import { InputBehaviour } from "../../__internal__/input-behaviour";
 import { filterStyledSystemMarginProps } from "../../style/utils";
-
 import InputIconToggle from "../../__internal__/input-icon-toggle";
-
 import guid from "../../utils/helpers/guid/guid";
 import StyledTextarea from "./textarea.style";
+import LocaleContext from "../../__internal__/i18n-context";
 
-const i18nNumberOpts = { precision: 0 };
+const getFormatNumber = (value, locale) =>
+  new Intl.NumberFormat(locale).format(value);
 
 const marginPropTypes = filterStyledSystemMarginProps(
   styledSystemPropTypes.space
@@ -29,6 +28,8 @@ const filterOutSpacingProps = (obj) =>
   );
 
 class Textarea extends React.Component {
+  static contextType = LocaleContext;
+
   // Minimum height of the textarea
   minHeight = 0;
 
@@ -113,7 +114,6 @@ class Textarea extends React.Component {
   get characterCount() {
     const value = this.props.value || "";
     const { characterLimit, warnOverLimit } = this.props;
-
     if (!characterLimit) {
       return null;
     }
@@ -121,8 +121,8 @@ class Textarea extends React.Component {
     return (
       <CharacterCount
         isOverLimit={this.overLimit && warnOverLimit}
-        value={I18n.toNumber(value.length, i18nNumberOpts)}
-        limit={I18n.toNumber(characterLimit, i18nNumberOpts)}
+        value={getFormatNumber(value.length, this.context.locale())}
+        limit={getFormatNumber(characterLimit, this.context.locale())}
         data-element="character-limit"
       />
     );