+++ src/components/tile-select/tile-select.component.js
@@ -1,8 +1,9 @@
 import PropTypes from "prop-types";
 import React, { useState, useEffect } from "react";
-import I18n from "i18n-js";
 import styledSystemPropTypes from "@styled-system/prop-types";
+
 import tagComponent from "../../utils/helpers/tags/tags";
+import useLocale from "../../hooks/__internal__/useLocale";
 import Button from "../button";
 
 import {
@@ -44,8 +45,8 @@ const TileSelect = ({
   footer,
   ...rest
 }) => {
+  const l = useLocale();
   const [hasFocus, setHasFocus] = useState(false);
-
   const handleDeselect = () =>
     onChange({
       target: {
@@ -67,7 +68,7 @@ const TileSelect = ({
           disabled={disabled}
           onClick={handleDeselect}
         >
-          {I18n.t("tileSelect.deselect", { defaultValue: "Deselect" })}
+          {l.tileSelect.deselect()}
         </Button>
       )
     );